MCRefiner Lmin and probability modification
Lmin is a parameter that controls the minimum length of an aligned block
in the subunits, similar to the fragment length of the CE alignment. The
default is set to 4 residues.

The probability is normalized to approach 0 when closer to maxIter. This
change allows faster convergence and a proportional optimization time in
function of the protein length.